+++ tests/test_query.py
@@ -28,21 +28,25 @@ def test_positional_mask():
 
 def test_population_mask():
     data = pd.DataFrame(range(3))
-    queries, mask = _circuit_mask(data, "default", {"population": "default", "other": "val"})
+    queries, mask = _circuit_mask(
+        data, "default", "biophysical", {"population": "default", "other": "val"}
+    )
     assert queries == {"other": "val"}
     npt.assert_array_equal(mask, [True, True, True])
 
-    queries, mask = _circuit_mask(data, "default", {"population": "unknown", "other": "val"})
+    queries, mask = _circuit_mask(
+        data, "default", "biophysical", {"population": "unknown", "other": "val"}
+    )
     assert queries == {"other": "val"}
     npt.assert_array_equal(mask, [False, False, False])
 
     queries, mask = _circuit_mask(
-        data, "default", {"population": "default", "node_id": [2], "other": "val"}
+        data, "default", "biophysical", {"population": "default", "node_id": [2], "other": "val"}
     )
     assert queries == {"other": "val"}
     npt.assert_array_equal(mask, [False, False, True])
 
-    queries, mask = _circuit_mask(data, "default", {"other": "val"})
+    queries, mask = _circuit_mask(data, "default", "biophysical", {"other": "val"})
     assert queries == {"other": "val"}
     npt.assert_array_equal(mask, [True, True, True])
 
@@ -105,28 +109,28 @@ def test_resolve_ids():
     data = pd.DataFrame(
         [[1, 0.4, "seven"], [2, 0.5, "eight"], [3, 0.6, "nine"]], columns=["int", "float", "str"]
     )
-    assert [False, True, False] == resolve_ids(data, "", {"str": "eight"}).tolist()
-    assert [False, False, False] == resolve_ids(data, "", {"int": 1, "str": "eight"}).tolist()
+    assert [False, True, False] == resolve_ids(data, "", "", {"str": "eight"}).tolist()
+    assert [False, False, False] == resolve_ids(data, "", "", {"int": 1, "str": "eight"}).tolist()
     assert [True, True, False] == resolve_ids(
-        data, "", {"$or": [{"str": "seven"}, {"float": (0.5, 0.5)}]}
+        data, "", "", {"$or": [{"str": "seven"}, {"float": (0.5, 0.5)}]}
     ).tolist()
     assert [False, False, True] == resolve_ids(
-        data, "", {"$and": [{"str": "nine"}, {"int": 3}]}
+        data, "", "", {"$and": [{"str": "nine"}, {"int": 3}]}
     ).tolist()
     assert [False, False, True] == resolve_ids(
-        data, "", {"$and": [{"str": "nine"}, {"$or": [{"int": 1}, {"int": 3}]}]}
+        data, "", "", {"$and": [{"str": "nine"}, {"$or": [{"int": 1}, {"int": 3}]}]}
     ).tolist()
     assert [False, True, True] == resolve_ids(
-        data, "", {"$or": [{"float": (0.59, 0.61)}, {"$and": [{"str": "eight"}, {"int": 2}]}]}
+        data, "", "", {"$or": [{"float": (0.59, 0.61)}, {"$and": [{"str": "eight"}, {"int": 2}]}]}
     ).tolist()
     assert [True, False, True] == resolve_ids(
-        data, "", {"$or": [{"node_id": 0}, {"edge_id": 2}]}
+        data, "", "", {"$or": [{"node_id": 0}, {"edge_id": 2}]}
     ).tolist()
-    assert [False, False, False] == resolve_ids(data, "", {"$or": []}).tolist()
-    assert [True, True, True] == resolve_ids(data, "", {"$and": []}).tolist()
+    assert [False, False, False] == resolve_ids(data, "", "", {"$or": []}).tolist()
+    assert [True, True, True] == resolve_ids(data, "", "", {"$and": []}).tolist()
 
     with pytest.raises(BluepySnapError) as e:
-        resolve_ids(data, "", {"str": {"$regex": "*.some", "edge_id": 2}})
+        resolve_ids(data, "", "", {"str": {"$regex": "*.some", "edge_id": 2}})
     assert "Value operators can't be used with plain values" in e.value.args[0]